refactor: sprint 42.1 + 42.2 — scancodeio package + transport extraction
Audit item 1.2 — split the 444-line libinv/services/scancodeio_client.py
into a focused package.

Sprint 42.1 (row 34) — package conversion:
- libinv/services/scancodeio/__init__.py: package body holding the
  ScancodeioClient façade and the symbols it composes.
- libinv/services/scancodeio_client.py: 5-line backwards-compat shim
  (re-exports * from libinv.services.scancodeio so every existing
  `from libinv.services.scancodeio_client import X` keeps working).

Sprint 42.2 (row 35) — transport extraction:
- libinv/services/scancodeio/transport.py: Session creation + retry/
  backoff logic + _request_json. The latent mypy return-value error
  on _request_json's narrowing (flagged in CHANGELOG as still-open
  Sprint 23 item) is now annotated correctly.

Verification: 109 passed, 1 pre-existing failure unchanged
(test_5xx_logs_and_raises — flake unrelated to this restructure),
mypy clean on libinv/services/scancodeio/. Identity assertion
ScancodeioClient is libinv.services.scancodeio.ScancodeioClient passes
across both import paths.
